# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POS Kafe Multi-Cabang API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /branches:
    get:
      operationId: getBranches
      tags: [branches]
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    patch:
      operationId: updateBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Branch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # USERS / STAFF
...
  /kitchen/queue:
    get:
      operationId: getKitchenQueue
      tags: [kitchen]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: station
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Kitchen order queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KitchenOrder"
  /kitchen/items/{itemId}/status:
    patch:
      operationId: updateKitchenItemStatus
      tags: [kitchen]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KitchenItemStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"

  # PAYMENTS
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # REPORTS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        branchId:
          type: ["integer", "null"]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, name, email, role, branchId, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, manager, cashier, waiter, chef, warehouse]
        branchId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Branch:
      type: object
      required: [id, name, address, phone, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        taxRate:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name, address, phone]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        taxRate:
          type: number
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        taxRate:
          type: number
        isActive:
          type: boolean

    # MENU CATEGORY
...
    OrderItem:
      type: object
      required: [id, orderId, menuItemId, menuItemName, quantity, unitPrice, totalPrice, kitchenStatus, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemImage:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        kitchenStatus:
          type: string
          enum: [new, processing, ready, served]
        station:
          type: ["string", "null"]
        createdAt:
          type: string
...
    KitchenOrder:
      type: object
      required: [orderId, orderNumber, tableNumber, items, createdAt]
      properties:
        orderId:
          type: integer
        orderNumber:
          type: string
        tableNumber:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        priority:
          type: integer
    KitchenItemStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, processing, ready, served]

    # PAYMENT
...
    Customer:
      type: object
      required: [id, name, phone, loyaltyPoints, totalSpend, visitCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer
        totalSpend:
          type: number
        visitCount:
          type: integer
        membershipTier:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer

    # REPORTS
...